backend/requirements.txt
absl-py==2.3.1
asgiref==3.8.1
astunparse==1.6.3
audioread==3.0.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
decorator==5.2.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.14.0
hf-xet==1.1.5
huggingface-hub==0.33.2
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
keras==3.10.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.1.1
namex==0.1.0
networkx==3.5
numba==0.61.2
numpy==1.26.4
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
platformdirs==4.3.8
pooch==1.8.2
protobuf==4.25.8
psycopg2-binary==2.9.10
pycparser==2.22
pydub==0.25.1
Pygments==2.19.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
sqlparse==0.5.3
sympy==1.14.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.2.2
torchaudio==2.2.2
tqdm==4.67.1
transformers==4.53.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
